Add Post.py class
Created a class file por Posts entity.
