+++ explorer/Post.py
@@ -1,0 +1,6 @@
+class Post:
+    def __init__(self, post_id, caption, img_url, hahstag):
+        self.post_id = post_id
+        self.caption = caption
+        self.hashtag = hashtag
+        self.pciture_url = pciture_url
